+++ Projects/Profile/Sources/Store/ProfileAction.swift
@@ -1,9 +1,25 @@
 import Presentation
 
-public struct ProfileAction: ActionProtocol, Hashable {
-    
+public enum ProfileAction: ActionProtocol {
+    case fetchProfileState
+    case openProfile
+    case openCharity
+    case openPayment
+    case openEuroBonus
+    case openChangeEuroBonus
+    case dismissChangeEuroBonus
+    case openSuccessChangeEuroBonus
+    case openFreeTextChat
+    case openAppInformation
+    case openAppSettings(animated: Bool)
+    case setMember(id: String, name: String, email: String, phone: String?)
+    case setMemberEmail(email: String)
+    case setMemberPhone(phone: String)
+    case setEurobonusNumber(partnerData: PartnerData?)
+    case fetchProfileStateCompleted
+    case updateEurobonusNumber(number: String)
+    case setOpenAppSettings(to: Bool)
 }
-
 
 public enum ProfileLoadingAction: LoadingProtocol {
     case loading
